Prevent overwriting of existing file
If the remediated name for a file matches the name of another file, the latter file
would be overwritten when the first file is renamed. A check was added to make sure
this does not happen. Instead, an error message will be logged, and the file name
will remain unchanged; the name will need to be remediated by hand.

+++ remediate_filepaths.py
@@ -40,26 +40,41 @@ def log_processing_msg(offset, type_arg, name):
     file_logger.info(offset + 'Processing ' + type_arg + ": '" + name + "'")
 
 def log_renamed_msg(offset, type_arg, old_name, new_name):
-    file_logger.info(offset + 'Renamed ' + type_arg + ": '" +
-                     old_name + "' to: '" +
-                     new_name + "'")
+    file_logger.info(offset + 'Renamed ' + type_arg)
+    file_logger.info(offset + "old: '" + old_name + "'")
+    file_logger.info(offset + "new: '" + new_name + "'")
+
+def log_name_exists_error(offset, type_arg, old_name, new_name):
+    file_logger.info(offset + '**** ERROR (BEGIN) **** -- ' + type_arg +
+                     " with new name already exists, so won't rename original " + type_arg + '.')
+    file_logger.info(offset + "old: '" + old_name + "'")
+    file_logger.info(offset + "new: '" + new_name + "'")
+    file_logger.info(offset + '**** ERROR (END) **** -- See above info')
 
 def process_dir(target_dir):
+    file_logger.info('Starting dir for script:')
+    file_logger.info("'" + target_dir + "'")
+    file_logger.info('All current dir paths in the log will be relative to the above path')
     # offset is used when logging output, to help with the formating
     offset = '  '
+    no_offset = ''
     # os.walk takes a topdown parameter which defaults to true.
     # Need to set it to true so it goes bottom-up.
     for dirpath, subdirs, files in os.walk(target_dir, topdown=False):
-        file_logger.info("Currently in dir: '" + dirpath + "'")
+        file_logger.info('Current dir: ')
+        file_logger.info("'" + os.path.relpath(dirpath,target_dir) + "'")
         # First, handle subdirs in current directory
         for subdir in subdirs:
             if args.verbose:
                 log_processing_msg(offset, 'subdir', subdir)
             remediated_name = remediate_basename_dirname(subdir)
             if remediated_name != subdir:
-                os.rename(os.path.join(dirpath, subdir),
-                          os.path.join(dirpath, remediated_name))
-                log_renamed_msg(offset, 'subdir', subdir, remediated_name)
+                if os.path.lexists(os.path.join(dirpath,remediated_name)):
+                    log_name_exists_error(no_offset, 'subdir', subdir, remediated_name)
+                else:
+                    os.rename(os.path.join(dirpath, subdir),
+                              os.path.join(dirpath, remediated_name))
+                    log_renamed_msg(offset, 'subdir', subdir, remediated_name)
         # Second, handle files in current directory
         for filename in files:
             if args.verbose:
@@ -69,9 +84,12 @@ def process_dir(target_dir):
             remediated_basename = remediate_basename_dirname(basename)
             remediated_extension = remediate_extension(extension) if extension else extension
             if remediated_basename != basename or remediated_extension != extension:
-                os.rename(os.path.join(dirpath, filename),
-                          os.path.join(dirpath, remediated_basename + remediated_extension))
-                log_renamed_msg(offset, 'file', filename, remediated_basename + remediated_extension)
+                if os.path.lexists(os.path.join(dirpath,remediated_basename + remediated_extension)):
+                    log_name_exists_error(no_offset, 'file', filename, remediated_basename + remediated_extension)
+                else:
+                    os.rename(os.path.join(dirpath, filename),
+                              os.path.join(dirpath, remediated_basename + remediated_extension))
+                    log_renamed_msg(offset, 'file', filename, remediated_basename + remediated_extension)
 
 def remediate_basename_dirname(str_arg):
     # first, handle non-ASCII characters
